Keep key points and add totally blank.
New feature:
1. Fixed bug of channel coding caused by buffer type error. Change the parameters of coding buffer from uint8_t * to uint16_t * in Interleave(), degray_code() and modulatation().
2. Add blank_position_cal() function to keep the key points in a symbol. 10% duration for start and end, 20% duration for jump point.
3. Add totally blank symbol in packet.
3. Add clear 0 of parameters( chip_position ) at the end of symbol transmitted.
4.

+++ code/SX1276/LoRa_Channel_Coding/Gray_decode.c
@@ -17,11 +17,11 @@ inline long mod(long a, long b)
 	return result_mod;
 }
 
-uint8_t *Gray_Decoder(uint8_t cr, uint8_t sf, char *input_str, uint8_t *input, uint8_t ninput_items, uint8_t *noutput_items)
+uint16_t *Gray_Decoder(uint8_t cr, uint8_t sf, char *input_str, uint16_t *input, uint8_t ninput_items, uint8_t *noutput_items)
 {
-	uint8_t *output;
+	uint16_t *output;
 	
-	output = malloc(ninput_items*sizeof(uint8_t));
+	output = malloc(ninput_items*sizeof(uint16_t));
 	
 	#ifdef DEBUG
 	for(int i=0;i<ninput_items;i++)
