+++ code/SX1276/LoRa_Channel_Coding/Interleaver.h
@@ -5,6 +5,6 @@
 #include <stdbool.h>
 
 long mod(long a, long b);
-uint8_t *Interleaver(uint8_t cr, uint8_t sf, char *input_str, uint8_t *input, uint8_t ninput_items, uint8_t *noutput_items);
+uint16_t *Interleaver(uint8_t cr, uint8_t sf, char *input_str, uint8_t *input, uint8_t ninput_items, uint8_t *noutput_items);
 
 #endif
