+++ code/SX1276/LoRa_Channel_Coding/Modulation.h
@@ -3,6 +3,6 @@
 
 #include "stm32l4xx.h"
 
-int *Modulation(uint8_t cr, uint8_t sf, uint32_t bw, uint8_t *input, uint8_t ninput_items, uint8_t *noutput_items);
+int *Modulation(uint8_t cr, uint8_t sf, uint32_t bw, uint16_t *input, uint8_t ninput_items, uint8_t *noutput_items);
 
 #endif
